requirements.txt
streamlit
pandas
pymongo
gTTS
python-dotenv
geocoder